made all fields compatible with long values

+++ src/main/java/de/blackforestsolutions/apiservice/service/communicationservice/VBBApiServiceImpl.java
@@ -39,7 +39,7 @@ public class VBBApiServiceImpl implements VBBApiService {
         return price.build();
     }
 
-    private static Map<PriceCategory, BigDecimal> convertPriceToPriceWithComma(int price) {
+    private static Map<PriceCategory, BigDecimal> convertPriceToPriceWithComma(long price) {
         StringBuilder sb = new StringBuilder(String.valueOf(price));
         int length = (int) Math.log10(price) + 1;
         sb.insert(length - 2, ".");
